requirements.txt
langchain-community
langchain-text-splitters
langchain-groq
langgraph
chromadb
python-dotenv